Update Sidebar.jsx
The box that includes the profile image, overlaps with the other lists in the sidebar. But for User experience, I was able to make them work independently

Here are the changes I made:
I made the Box position sticky rather than absolute so that the other list could flow independently without colliding with the profile box. And added some flex layout on the other boxes

I have the explanations with Images in the issues section.